Adding status to responses

+++ ecomerce_rest/apps/users/api/api.py
@@ -1,3 +1,4 @@
+from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.decorators import api_view
@@ -10,34 +11,35 @@ def user_api_view(request):
     if request.method == 'GET':
         users = User.objects.all()
         users_serializers = UserSerializer(users, many=True)
-        return Response(users_serializers.data)
+        return Response(users_serializers.data, status=status.HTTP_200_OK)
 
     elif request.method == 'POST':
         user_serializer = UserSerializer(data=request.data)
         if user_serializer.is_valid():
             user_serializer.save()
-            return Response(user_serializer.data)
+            return Response(user_serializer.data, status=status.HTTP_201_CREATED)
         else:
-            return Response(user_serializer.errors)
+            return Response(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET','PUT','DELETE'])
 def user_detail_api_view(request, pk=None):
+    user = User.objects.filter(id=pk).first()
+    if user:
+        if request.method == 'GET':
+            user_serializer = UserSerializer(user)
+            return Response(user_serializer.data, status=status.HTTP_200_OK)
 
-    if request.method == 'GET':
-        user = User.objects.filter(id=pk).first()
-        user_serializer = UserSerializer(user)
-        return Response(user_serializer.data)
+        elif request.method == 'PUT':
 
-    elif request.method == 'PUT':
-        user = User.objects.filter(id=pk).first()
-        user_serializer = UserSerializer(user, data=request.data)
-        if user_serializer.is_valid():
-            user_serializer.save()
-            return Response(user_serializer.data)
-        else:
-            return Response(user_serializer.errors)
+            user_serializer = UserSerializer(user, data=request.data)
+            if user_serializer.is_valid():
+                user_serializer.save()
+                return Response(user_serializer.data, status=status.HTTP_200_OK)
+            else:
+                return Response(user_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    elif request.method == 'DELETE':
-        user = User.objects.filter(id=pk).first()
-        user.delete()
-        return Response('Eliminado')
+        elif request.method == 'DELETE':
+            user.delete()
+            return Response({'message':'Eliminado'}, status=status.HTTP_200_OK)
+    else:
+        return Response({'message':'Usuario no encontrado'}, status=status.HTTP_400_BAD_REQUEST)
